+++ ARRAYS/Stack as Linked list
