+++ ARRAYS/majority_element.cpp
