+++ ARRAYS/unionofarray.cpp
