+++ DYNAMIC PROGRAMMING/dot_product.cpp
