+++ STACKS/stack_longest_valid_substring.cpp
